Rewind: Refresh the Activity Log after a Rewind operation
After a Rewind operation finishes then the entire list of
events in the Activity Log could be stale. The `isDiscarded`
property is affected by all future rewinds.

Therefore in this patch I'm attempting to refresh all of the
previously-fetched Activity Log events once we transition from
_Rewinding_ to _Not Rewinding_. This should clear up the stale
state.

This system relies on the fact that the Rewind API itself is going to
poll for updates whenever the Rewind status is `running`. That is, it
will automatically keep itself up-to-date until the Rewind finishes,
upon which we can make a decision whether or not to refresh the
Activity Log.

Future work needs to additionally make sure we pull in new events
because the very act of Rewinding creates a completion event. Those are
currently being overlooked.

Additionally, the complexity here is somewhat increased due to the way
that the `ActivityQueryManager` stores its own state. We are introducing
code that is strongly couples to several different implementation
details in order to perform the data refreshses. In the future it would
be nice if we had a clearer way to do this; I could be missing important
semantics of the `QueryManager` though.