Replace tightly-coupled requester with better one and delay updates
First of all I just realized that the query manager doesn't force us
to stick to working with the queries used when requesting the objects.
This is great. Because of that I'm changing the code here to find the
oldest activity that we have in local state and then refetching all
requests from then until the present time. This dramatically simplifies
what we're doing in this PR.

Sadly though when we request the updates to the Activity Log th backend
services havnen't updated as well. I'm not sure if it's ElasticSearch
lag or som caching or any other reason; all I know is that we're getting
stale updates from the server. To counter this I'm delaying the fetch of
the Activity Log entries to mitigate this data race. A proper solution
would eliminate the race but this should get us far enough for now.

After this commit I hope to update the data layer handler such that it
will update the local store with recomputed `isDiscarded` properties.
It's primarily those that are important here anyway. We'll keep the
delayed fetch but provide an instant approximation of how the update
will look.